[hotfix] Refactor NeverCompleteFuture into a reusable utility
